유저랑 그룹 프로바이더는 어느정도.. 테스크가 너무 헬이라 막힘

+++ lib/screens/group/groupmanagedialog.dart
@@ -7,7 +7,8 @@ class GroupManageDialog extends StatefulWidget {
   final String user_name;
   final String groupId;
 
-  const GroupManageDialog({super.key, 
+  const GroupManageDialog({
+    super.key,
     required this.user_id,
     required this.user_name,
     required this.groupId,
@@ -22,8 +23,6 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
   List<Map<String, dynamic>> tasks = [];
   Map<String, dynamic> userTasksState = {};
 
-
-
   @override
   void initState() {
     super.initState();
@@ -31,29 +30,35 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
   }
 
   Future<void> _fetchUserTasks() async {
-    QuerySnapshot taskSnapshot = await firestore
-        .collection('tasks')
-        .where('assigned_users', arrayContains: widget.user_id)
-        .get();
+    QuerySnapshot taskSnapshot =
+        await firestore
+            .collection('tasks')
+            .where('assigned_users', arrayContains: widget.user_id)
+            .get();
 
-    for (var doc in taskSnapshot.docs) {if(doc['managed_by']!=''){
-      Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
-      int userIndex = (data['assigned_users'] as List).indexOf(widget.user_id);
-      if (userIndex != -1) {
-        tasks.add({
-          'id': doc.id,
-          'name': data['task_name'],
-          'state': data['state'][userIndex],
-          'assigned_users': data['assigned_users'],
-        });
-      }}
+    for (var doc in taskSnapshot.docs) {
+      if (doc['managed_by'] != '') {
+        Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
+        int userIndex = (data['assigned_users'] as List).indexOf(
+          widget.user_id,
+        );
+        if (userIndex != -1) {
+          tasks.add({
+            'id': doc.id,
+            'name': data['task_name'],
+            'state': data['state'][userIndex],
+            'assigned_users': data['assigned_users'],
+          });
+        }
+      }
     }
 
     setState(() {});
   }
 
   void _updateTaskState(String taskId, int newState, int userIndex) async {
-    DocumentSnapshot taskDoc = await firestore.collection('tasks').doc(taskId).get();
+    DocumentSnapshot taskDoc =
+        await firestore.collection('tasks').doc(taskId).get();
     if (!taskDoc.exists) return;
 
     Map<String, dynamic> taskData = taskDoc.data() as Map<String, dynamic>;
@@ -64,7 +69,6 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
       firestore.collection('tasks').doc(taskId).update({'state': states});
     }
   }
-
 
   void _forceRemoveUser(BuildContext context) {
     showDialog(
@@ -93,19 +97,18 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
       },
     );
   }
-  Future<void> leaveGroup(String? uid) async {
 
+  Future<void> leaveGroup(String? uid) async {
     if (uid == null) return;
 
-    final userDoc = await FirebaseFirestore.instance
-        .collection('users')
-        .doc(uid)
-        .get();
+    final userDoc =
+        await FirebaseFirestore.instance.collection('users').doc(uid).get();
     final user_group_id = userDoc.data()?['user_group_id'];
-    final groupDoc = await FirebaseFirestore.instance
-        .collection('groups')
-        .doc(user_group_id)
-        .get();
+    final groupDoc =
+        await FirebaseFirestore.instance
+            .collection('groups')
+            .doc(user_group_id)
+            .get();
     final group_manager_id = groupDoc.data()?['group_manager'];
 
     if (user_group_id != null) {
@@ -114,23 +117,22 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
           .collection('groups')
           .doc(user_group_id)
           .update({
-        'group_users': FieldValue.arrayRemove([uid])
-      });
+            'group_users': FieldValue.arrayRemove([uid]),
+          });
 
       // 사용자의 user_group_id 필드 비우기
-      await FirebaseFirestore.instance
-          .collection('users')
-          .doc(uid)
-          .update({'user_group_id': ''});
+      await FirebaseFirestore.instance.collection('users').doc(uid).update({
+        'user_group_id': '',
+      });
 
-      // 관리하는 작업에서 사용자 제거
-      final tasksQuerySnapshot = await FirebaseFirestore.instance
-          .collection('tasks')
-          .where('managed_by', isNotEqualTo: uid)
-          .get();
+      // 작업에서 사용자 제거(자기가 관리 안하는 작업)
+      final tasksQuerySnapshot =
+          await FirebaseFirestore.instance
+              .collection('tasks')
+              .where('managed_by', isNotEqualTo: uid)
+              .get();
 
       for (var task in tasksQuerySnapshot.docs) {
-
         var assignedUsers = List.from(task['assigned_users']);
         if (assignedUsers.contains(uid)) {
           int userIndex = assignedUsers.indexOf(uid);
@@ -146,23 +148,25 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
             await task.reference.update({
               'assigned_users': assignedUsers,
               'state': states,
-              'alert': alerts
+              'alert': alerts,
+              'managed_by': group_manager_id,
             });
           }
-
         }
       }
 
       //내가 관리하는 업무를 관리자에게 떠넘기기
-      final tasksQuerySnapshot2 = await FirebaseFirestore.instance
-          .collection('tasks')
-          .where('managed_by', isEqualTo: uid)
-          .get();
+      final tasksQuerySnapshot2 =
+          await FirebaseFirestore.instance
+              .collection('tasks')
+              .where('managed_by', isEqualTo: uid)
+              .get();
 
-      final groupDoc = await FirebaseFirestore.instance
-          .collection('groups')
-          .doc(user_group_id)
-          .get();
+      final groupDoc =
+          await FirebaseFirestore.instance
+              .collection('groups')
+              .doc(user_group_id)
+              .get();
 
       for (var doc in tasksQuerySnapshot2.docs) {
         List<dynamic> assignedUsers = doc['assigned_users'] ?? [];
@@ -181,21 +185,23 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
           'assigned_users': assignedUsers,
           'managed_by': task_manager,
           'state': states,
-          'alert': alerts
+          'alert': alerts,
         });
       }
 
       // 작성한 게시물 삭제
-      final postsQuerySnapshot = await FirebaseFirestore.instance
-          .collection('posts')
-          .where('made_by', isEqualTo: uid)
-          .get();
+      final postsQuerySnapshot =
+          await FirebaseFirestore.instance
+              .collection('posts')
+              .where('made_by', isEqualTo: uid)
+              .get();
 
       for (var doc in postsQuerySnapshot.docs) {
-        var commentsSnapshot = await FirebaseFirestore.instance
-            .collection('comments')
-            .where('post_id', isEqualTo: doc.id)
-            .get();
+        var commentsSnapshot =
+            await FirebaseFirestore.instance
+                .collection('comments')
+                .where('post_id', isEqualTo: doc.id)
+                .get();
         for (var docu in commentsSnapshot.docs) {
           await docu.reference.delete();
         }
@@ -205,28 +211,29 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
         List<dynamic> fileUrls = doc['fileUrl'] ?? [];
         List<dynamic> imageUrls = doc['imageUrl'] ?? [];
 
-        for (var fileUrl in fileUrls) {
-          if (fileUrl != null) {
-            var fileRef = storage.refFromURL(fileUrl);
-            await fileRef.delete();
+        if (fileUrls.isNotEmpty) {
+          for (var fileUrl in fileUrls) {
+            if (fileUrl != null) {
+              var fileRef = storage.refFromURL(fileUrl);
+              await fileRef.delete();
+            }
           }
         }
 
-        for (var imageUrl in imageUrls) {
-          if (imageUrl != null) {
-            var imageRef = storage.refFromURL(imageUrl);
-            await imageRef.delete();
+        if (imageUrls.isNotEmpty) {
+          for (var imageUrl in imageUrls) {
+            if (imageUrl != null) {
+              var imageRef = storage.refFromURL(imageUrl);
+              await imageRef.delete();
+            }
           }
         }
         await doc.reference.delete();
       }
 
       // 그룹 탈퇴 처리 후 로직 (예: 홈 화면으로 이동)
-
-
     }
   }
-
 
   String taskStateToString(int state) {
     switch (state) {
@@ -260,40 +267,48 @@ class _GroupManageDialogState extends State<GroupManageDialog> {
       title: Text(widget.user_name),
       content: SingleChildScrollView(
         child: ListBody(
-          children: tasks.map((task) {
-            return ListTile(
-              title: Text(task['name']),
-              trailing: DropdownButton<String>(
-                value: taskStateToString(task['state']),
-                items: <String>['시작 전', '진행중', '완료'].map((String value) {
-                  return DropdownMenuItem<String>(
-                    value: value,
-                    child: Text(value),
-                  );
-                }).toList(),
-                onChanged: (String? newValue) {
-                  if (newValue != null) {
-                    int newState = taskStateFromString(newValue);
-                    int userIndex = task['assigned_users'].indexOf(widget.user_id);
+          children:
+              tasks.map((task) {
+                return ListTile(
+                  title: Text(task['name']),
+                  trailing: DropdownButton<String>(
+                    value: taskStateToString(task['state']),
+                    items:
+                        <String>['시작 전', '진행중', '완료'].map((String value) {
+                          return DropdownMenuItem<String>(
+                            value: value,
+                            child: Text(value),
+                          );
+                        }).toList(),
+                    onChanged: (String? newValue) {
+                      if (newValue != null) {
+                        int newState = taskStateFromString(newValue);
+                        int userIndex = task['assigned_users'].indexOf(
+                          widget.user_id,
+                        );
 
-                    if (userIndex != -1) {
-                      setState(() {
-                        // UI 업데이트를 위해 tasks 리스트의 해당 태스크 상태 업데이트
-                        int taskIndex = tasks.indexWhere((t) => t['id'] == task['id']);
-                        if (taskIndex != -1) {
-                          tasks[taskIndex]['state'] = newState;
+                        if (userIndex != -1) {
+                          setState(() {
+                            // UI 업데이트를 위해 tasks 리스트의 해당 태스크 상태 업데이트
+                            int taskIndex = tasks.indexWhere(
+                              (t) => t['id'] == task['id'],
+                            );
+                            if (taskIndex != -1) {
+                              tasks[taskIndex]['state'] = newState;
+                            }
+
+                            _updateTaskState(
+                              task['id'],
+                              newState,
+                              userIndex,
+                            ); // Firestore 업데이트
+                          });
                         }
-
-                        _updateTaskState(task['id'], newState, userIndex); // Firestore 업데이트
-                      });
-                    }
-                  }
-                },
-
-
-              ),
-            );
-          }).toList(),
+                      }
+                    },
+                  ),
+                );
+              }).toList(),
         ),
       ),
       actions: <Widget>[
